Move the DotSlash build process out of scm/hg/setup.py and into hgbuild.
Summary:
As requested by DurhamG, bundle DotSlash in the same way that scmtelemetry
is bundled rather than as directly part of the Mercurial package.
This approach makes it possible for DotSlash to leverage Hg's contbuild
while still being in its own package.

To implement this, I copypasta'd scmtelemetry's setup and modified it
for DotSlash, as appropriate.

Although I tried to port the Windows configuration over in addition to
Mac and Linux, I'm going to disable it for now because I want to limit
how many things I'm going to try to get working on the initial rollout.

Reviewed By: singhsrb

Differential Revision: D13950939

fbshipit-source-id: 5d8393d756e23f1ac228c7315a4b95a36288ac3a