rename CheckoutConfig.path to backing_repo
Summary:
This renames the CheckoutConfig `path` member variable to `backing_repo`.
The value in this variable is actually the path to the source control backing
repository, and not the path to the Eden checkout, which was confusing.  (The
CLI tests were actually setting this up incorrectly in `create_test_mount()`,
but until now nothing in the tests had been using it, so we hadn't noticed
this problem earlier.)

I also changed `CheckoutConfig` to be a `typing.NamedTuple` as part of this
change.

Reviewed By: strager

Differential Revision: D13935006

fbshipit-source-id: 43b6ba4efa34f8c6fb8264de8bb31a33a6d73036