test-fb-hgext-fastannotate-remotefile: remove the test
Summary:
The test was about the connection handling between remotefilelog and
fastannotate. It uses named branch to test certain code paths, which
is not trivial to migrate off.

We'll probably rework the protocol used here and maybe do something very
different regarding on linelog with Mononoke. Therefore just remove the
test for now.

Reviewed By: singhsrb

Differential Revision: D13934435

fbshipit-source-id: 9ba0b760ba17efd3ede6de7612e6532f3d13a4f8