fix false positives in "eden doctor" with older edenfs versions
Summary:
This fixes a bug where "eden doctor" would incorrectly report that some
checkouts were not mounted when the version of `edenfs` was too old.  It would
think it needed to remount the checkouts even though they were in fact working
correctly.

Older versions of `edenfs` did not return a `state` field in the
`listMounts()` reply.  This caused the newer `eden doctor` logic in D13526754
to incorrectly think the mounts were not running.  This fixes the code to
assume the state is `RUNNING` in this case.

Reviewed By: wez

Differential Revision: D14017363

fbshipit-source-id: 020135f386aad84b1a92f2699a2706ba73141ec7