Add flag to toggle skiplist population
Summary:
Populating skiplist at startup takes up to 3 minutes when not compiling in prod mode. This allows developers to turn this off and just use an empty skiplist (this only affects historical queries, and those can still be tested using recent commits)

The reason for negating the flag is to not hide core functionality behind a flag.

Reviewed By: StanislavGlebik

Differential Revision: D13986936

fbshipit-source-id: 053ecbd27aeeaa97f03194933e1d904a3e60ffcc