make fb303.FacebookService derive from fb303_core.BaseService
Summary:
Update the `fb303.FacebookService` type to derive from the new
`fb303_core.BaseService` definition.  This is in preparation to begin moving
most of the fb303 methods to `BaseService`.

Unfortunately the Rust thrift compiler does not handle service inheritance
very well, so all existing rust thrift servers need to be updated to handle
this change.

Reviewed By: wez

Differential Revision: D13683717

fbshipit-source-id: c6ba8dcbadd05d674b7a75ea531500268e4bc3f2