mononoke: Log read requests for all blobstores
Summary:
Previously we didn't log any errors for reads as we were cancelling
outstanding futures for read requests as soon as one completed.

Modify this behaviour so that every time we decide to sample a request, we
allow all blobstore gets() to complete and record the data. As a result
of this, we will also record the errors.

Reviewed By: StanislavGlebik

Differential Revision: D14022739

fbshipit-source-id: 62b878a08b5c4f6a2ed9efe71b2f97cf66c58256