test-clone-update-order: remove the test
Summary:
The test uses named branches extensively, partically because there are multiple
ways to specify where to update:

  hg clone URL#NAME1 -b NAME2 -u NAME3
              ^^^^^^    ^^^^^    ^^^^^
         branch name    |        |
           branch name, again!   |
                  revset, can also be a branch name

In the long run, we'll kill NAME1 (`#...` syntax) and NAME2 (`clone -b` flag).
So the (undesirable) complexity will be gone. Remove the test to make the
transition easier.

The `clone -u REV` feature is still tested in `test-clone.t`.

Reviewed By: singhsrb

Differential Revision: D13978578

fbshipit-source-id: 839ad39779c4526e4be48e133610bdb46f57fc93