test-graft: rewrite test cases to use smaller repo
Summary:
Rewrite "graft --force" and "empty graft" test cases to not depend on the
previous repo state. Remove the "same origin policy" test since eventually
we'll use mutation history to test origins and the code path here would be
rewritten.

This makes it easier to change the rest of the test.

Differential Revision: D14000100

fbshipit-source-id: 26002774c370a965a9bca157e08246c89aa5e274