hg: rename scratch to mkscratch and include in the hg build
Summary:
My goal is to have `mkscratch` get deployed with the mercurial
release on all platforms.   This is currently deployed via packman
as `scratch` on devservers, and that packman build should continue to
succeed because the TARGETS file in here is left unchanged.

That will allow both `scratch` and `mkscratch` to coexist for a while
until we've made it through a few mercurial releases, and then we can
either fixup existing users of `scratch` to use `mkscratch`, or re-do
the `fb-scratch` rpm so that it symlinks to `mkscratch`.

I don't really know if this diff is sufficient to include the binary
in the mercurial package; I'm cargo culting the `scm_daemon` stuff
and that seems to be included in the omnibus, so maybe this is it?

Reviewed By: quark-zju

Differential Revision: D14060864

fbshipit-source-id: 5fd74a877437d84fe7676819c2cefb902fb7002d