eden: reinstate treemanifest support on macos
Summary:
This got lost partly because the symlink/shipit bit
was a manual step when I did this first time around, and partly
because the mercurial code got refactored.

Restoring treemanifest restores `hg up` performance to saner
levels.

This is fairly straightforward, but now that we have !wez building this we've
found that my machine had lz4 installed in the default include path and that we
need to probe and find the correct one, so that is part of this diff also.

Reviewed By: simpkins

Differential Revision: D14078886

fbshipit-source-id: 971cd49757bbb7dcc484439e6d896698598a583b