mononoke: Allow a dynamic stat prefix for CountedBlobstore
Summary:
Previously, CountedBlobstore used many dynamic_timeseries in its
stats structure to allow for a dynamic key based upon blobstore name. Replace
this with a stats struct that allows a dynamic prefix. This simplfies the
stats code for CountedBlobstore and also allows for a String to be used
as a prefix.

Differential Revision: D14084622

fbshipit-source-id: 1c6bdbe63e5c4f6e9f0428c9b3cdfdbed475b877