tests: make test-fb-hgext-remotefilelog-corrupt-cache.t reliable
Summary:
This test has been extremely flaky in the last few weeks, causing builds to not
being released, or just sandcastle jobs failing randomly. The reason for the
flakyness was due to the mtime stored in the dirstate being -1 instead of the
real mtime of the file.

The reason for the mtime to be -1 is to fix a race where a file is modified in
the same second the dirstate is updated. This can happen when a transaction
spans less than a second. By forcing a transaction to always last at least a
second, we can avoid this issue and make the test stable.

Reviewed By: quark-zju

Differential Revision: D14141924

fbshipit-source-id: ce151cd65d930f8c745633f7575ea006748af9b9