Fix creation for empty RepoPath
Summary:
While working on TreeManifest I found that the code does not allow creating an empty `RepoPath`. I did not realize that splitting an empty string always results in an empty string. This is an edge case and we need to handle it appropriately.

Currently adding a hack where the components iterator function will skip empty values. We are going to add more iterators in a future revision and we should revisit this function then.

Reviewed By: quark-zju

Differential Revision: D14142927

fbshipit-source-id: 1be43d61913138e3fa50e02976e81f9ca38a74a7